requirements.txt
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
entrypoints==0.3
httpie==2.2.0
psycopg2==2.8.5
pylint==2.5.2
webencodings==0.5.1
whitenoise==5.1.0
gunicorn
PyJwt
django-push-notifications
django_cron
apscheduler
django-environ
fcm_django
django-allauth
drf-yasg
requests